requirements.txt
bleach==6.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
kaggle==1.6.3
pylab-sdk==1.7.2
python-slugify==8.0.1
pyyaml==6.0.1
requests==2.31.0
text-unidecode==1.3
tqdm==4.66.1
urllib3==2.1.0
webencodings==0.5.1
numpy 
pandas
scikit-learn
ray[serve]==2.9.2
kserve==0.12.0

